コードをさらに整理し、Boardクラスに含まれるメンバ関数を削減した。
